move App.js into the root of src
