+++ index.android.js
@@ -21,7 +21,7 @@
  */
 
 import { AppRegistry, YellowBox } from 'react-native'
-import App from './src/components/App'
+import App from './src/App'
 YellowBox.ignoreWarnings(['Warning: isMounted(...) is deprecated', 'Module RCTImageLoader'])
 
 AppRegistry.registerComponent('NativeSigner', () => App)
