+++ src/App.js
@@ -23,19 +23,19 @@ import { StyleSheet, AppState, Alert } from 'react-native'
 import { Provider } from 'react-redux'
 import { StackNavigator, TabNavigator, TabBarBottom } from 'react-navigation'
 import { Actions, Scene } from 'react-native-router-flux'
-import TabIcon from './TabIcon'
-import IconChooser from '../containers/IconChooser'
-import QrScanner from '../containers/QrScanner'
-import AccountList from '../containers/AccountList'
-import AccountNew from '../containers/AccountNew'
-import AccountDetails from '../containers/AccountDetails'
-import TxDetails from '../containers/TxDetails'
-import { AccountEnterPin, AccountChangePin, AccountSetPin, AccountConfirmPin } from '../containers/AccountPin'
-import { QrViewTransaction } from '../containers/QrView'
-import DataDetails from '../containers/DataDetails'
-import { loadAccounts, saveAccounts } from '../util/db'
-import { setAccounts } from '../actions/accounts'
-import store from '../util/store'
+import TabIcon from './components/TabIcon'
+import IconChooser from './containers/IconChooser'
+import QrScanner from './containers/QrScanner'
+import AccountList from './containers/AccountList'
+import AccountNew from './containers/AccountNew'
+import AccountDetails from './containers/AccountDetails'
+import TxDetails from './containers/TxDetails'
+import { AccountEnterPin, AccountChangePin, AccountSetPin, AccountConfirmPin } from './containers/AccountPin'
+import { QrViewTransaction } from './containers/QrView'
+import DataDetails from './containers/DataDetails'
+import { loadAccounts, saveAccounts } from './util/db'
+import { setAccounts } from './actions/accounts'
+import store from './util/store'
 
 loadAccounts().then(accounts => {
   store.dispatch(setAccounts(accounts))
