+++ src/screens/QrScanner.js
